fix case-sensitive inconsistency of name of Location header (#42)
Reporter found that AWS servers sometimes return the 'Location' header as 'location'.  Because the 'create' method of PBS MM API only returns the actual ID of the created element within the Location header, this provided unexpected errors if software wanted to get that ID to use further.  Making the preg_match regex that searches for Location case-insensitive was trivial.